requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
sentence-transformers
python-dotenv
pandas
matplotlib
